# server/docs/swagger/swagger.yaml
basePath: /
definitions:
...
  github_com_ix-pay_ixpay-pro_internal_domain_base_entity.Position:
    properties:
      createdAt:
        description: 创建时间
        type: string
      createdBy:
        description: 创建人 ID
        type: string
      description:
        description: 岗位描述
        type: string
      id:
        description: 岗位 ID
        type: string
      name:
        description: 岗位名称
        type: string
      sort:
        description: 排序
        type: integer
      status:
        description: 状态：1-正常，0-禁用
        type: integer
      updatedAt:
        description: 更新时间
        type: string
      updatedBy:
        description: 更新人 ID
        type: string
    type: object
...
  github_com_ix-pay_ixpay-pro_internal_dto_base_request.CreatePositionRequest:
    properties:
      description:
        description: 岗位描述
        type: string
      name:
        description: 岗位名称
        type: string
      sort:
        description: 排序，默认为 0
        type: integer
      status:
        description: 状态：1-正常，0-禁用，默认为 1
        type: integer
    required:
    - name
    type: object
...
  github_com_ix-pay_ixpay-pro_internal_dto_base_request.UpdatePositionRequest:
    properties:
      description:
        description: 岗位描述
        type: string
      id:
        description: 岗位 ID
        type: string
      name:
        description: 岗位名称
        type: string
      sort:
        description: 排序
        type: integer
      status:
        description: 状态：1-正常，0-禁用
        type: integer
    required:
    - id
    - name
    type: object
...
  github_com_ix-pay_ixpay-pro_internal_dto_base_response.PositionListResponse:
    properties:
      list:
        description: 岗位列表
        items:
          $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_dto_base_response.PositionResponse'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  github_com_ix-pay_ixpay-pro_internal_dto_base_response.PositionResponse:
    properties:
      code:
        description: 岗位编码
        type: string
      createdAt:
        description: 创建时间
        type: string
      description:
        description: 描述
        type: string
      id:
        description: 岗位 ID
        example: ""
        type: string
      name:
        description: 岗位名称
        type: string
      sort:
        description: 排序
        type: integer
      status:
        description: 状态：1-启用 0-禁用
        type: integer
      updatedAt:
        description: 更新时间
        type: string
    type: object
...
  github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
host: 127.0.0.1:8586
info:
  contact:
    email: support@example.com
    name: API Support
  description: 基于Gin和PostgreSQL的微信支付API服务
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 微信支付API服务
  version: "1.0"
paths:
...
  /api/admin/position:
    get:
      consumes:
      - application/json
      description: 获取岗位列表（支持分页和筛选）
      parameters:
      - description: 页码
        in: query
        name: page
        required: true
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        required: true
        type: integer
      - description: 状态 (0:禁用，1:启用)
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 岗位列表
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_dto_base_response.PositionListResponse'
                msg:
                  type: string
              type: object
        "401":
          description: 未授权
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取岗位列表
      tags:
      - 岗位管理
    post:
      consumes:
      - application/json
      description: 创建新的岗位
      parameters:
      - description: 岗位信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_dto_base_request.CreatePositionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_dto_base_response.PositionResponse'
                msg:
                  type: string
              type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 创建岗位
      tags:
      - 岗位管理
    put:
      consumes:
      - application/json
      description: 更新岗位信息
      parameters:
      - description: 岗位信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_dto_base_request.UpdatePositionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_dto_base_response.PositionResponse'
                msg:
                  type: string
              type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 更新岗位
      tags:
      - 岗位管理
  /api/admin/position/:id:
    get:
      consumes:
      - application/json
      description: 根据 ID 获取岗位详细信息
      parameters:
      - description: 岗位 ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 岗位详情
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_dto_base_response.PositionResponse'
                msg:
                  type: string
              type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取岗位详情
      tags:
      - 岗位管理
  /api/admin/position/{id}:
    delete:
      consumes:
      - application/json
      description: 删除岗位（管理员权限）
      parameters:
      - description: 岗位 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.Response'
            - properties:
                msg:
                  type: string
              type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 删除岗位
      tags:
      - 岗位管理
  /api/admin/position/all:
    get:
      consumes:
      - application/json
      description: 获取完整的岗位列表数据
      produces:
      - application/json
      responses:
        "200":
          description: 岗位列表
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_domain_base_entity.Position'
                  type: array
                msg:
                  type: string
              type: object
        "401":
          description: 未授权
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取所有岗位
      tags:
      - 岗位管理
...
securityDefinitions:
  BearerAuth:
    in: header [Bearer ]
    name: Authorization
    type: apiKey
swagger: "2.0"